requirements.txt
alabaster==0.7.7
amqp==1.4.8
anyjson==0.3.3
appnope==0.1.0
Babel==2.1.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
billiard==3.3.0.22
brewer2mpl==1.4.1
celery==3.1.19
certifi==2015.11.20.1
click==6.2
cycler==0.9.0
decorator==4.0.6
docutils==0.12
feedparser==5.2.1
Flask==0.10.1
Flask-SocketIO==2.0
Flask-Sockets==0.1
Flask-SQLAlchemy==2.1
functools32==3.2.3.post2
gevent==1.0.2
gevent-websocket==0.9.5
ggplot==0.6.8
gnureadline==6.3.3
greenlet==0.4.9
gunicorn==19.4.1
ipykernel==4.2.2
ipython==4.0.1
ipython-genutils==0.1.0
ipywidgets==4.1.1
iron-cache==0.3.2
iron-celery==0.4.5
iron-core==1.2.0
iron-mq==0.8
itsdangerous==0.24
Jinja2==2.7.3
joblib==0.9.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.1.0
jupyter-core==4.0.6
kombu==3.0.32
MarkupSafe==0.23
matplotlib==1.5.1
mechanize==0.2.5
memory-profiler==0.40
mistune==0.7.1
multiprocessing==2.6.2.1
nbconvert==4.1.0
nbformat==4.0.1
notebook==4.1.0
numpy==1.10.4
palettable==2.1.1
pandas==0.17.1
Paste==2.0.2
path.py==8.1.2
patsy==0.4.1
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
psutil==3.3.0
psycopg2==2.6.1
ptyprocess==0.5
Pygments==2.0.2
pyparsing==2.1.0
python-dateutil==2.4.2
python-engineio==0.8.6
python-socketio==0.9.1
pytz==2015.7
pyzmq==15.2.0
qtconsole==4.1.1
redis==2.10.5
requests==2.9.1
rq==0.5.6
scipy==0.16.1
selenium==2.48.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.3
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.0.11
statsmodels==0.6.1
terminado==0.6
tornado==4.3
traitlets==4.0.0
virtualenv==13.1.2
Werkzeug==0.10.4
wheel==0.26.0
